EIP OVN Controller: NAT only when default cluster network
..because UDN NAT to EIP will occur on br-ex.

+++ go-controller/pkg/ovn/base_network_controller_egressip.go
@@ -1615,7 +1615,7 @@ func (e *egressIPZoneController) addPodEgressIPAssignment(egressIPName string, s
 	}
 	var ops []ovsdb.Operation
 	if loadedEgressNode && isLocalZoneEgressNode {
-		if isOVNNetwork {
+		if e.IsDefault() && isOVNNetwork {
 			ops, err = e.createNATRuleOps(nil, podIPs, status, egressIPName)
 			if err != nil {
 				return fmt.Errorf("unable to create NAT rule ops for status: %v, err: %v", status, err)
@@ -1708,9 +1708,11 @@ func (e *egressIPZoneController) deletePodEgressIPAssignment(egressIPName string
 				return fmt.Errorf("unable to delete logical router static route ops %v, err: %v", status, err)
 			}
 		}
-		ops, err = e.deleteNATRuleOps(ops, podIPs, status, egressIPName)
-		if err != nil {
-			return fmt.Errorf("unable to delete NAT rule for status: %v, err: %v", status, err)
+		if e.IsDefault() {
+			ops, err = e.deleteNATRuleOps(ops, podIPs, status, egressIPName)
+			if err != nil {
+				return fmt.Errorf("unable to delete NAT rule for status: %v, err: %v", status, err)
+			}
 		}
 	}
 	_, err = libovsdbops.TransactAndCheck(e.nbClient, ops)
@@ -2025,7 +2027,7 @@ func (e *egressIPZoneController) deleteEgressIPStatusSetup(name string, status e
 		klog.Errorf("Unable to get next hop IP and therefore there could be stale logical route policies for Egress IP %s", status.EgressIP)
 	}
 
-	if loadedEgressNode && isLocalZoneEgressNode {
+	if e.IsDefault() && loadedEgressNode && isLocalZoneEgressNode {
 		routerName := e.GetNetworkScopedGWRouterName(status.Node)
 		natPred := func(nat *nbdb.NAT) bool {
 			// We should delete NATs only from the status.Node that was passed into this function
